+++ NEWS.md
@@ -1,6 +1,9 @@
 1.30.2 - 2016-mm-dd
 -------------------
 
+New features:
+ * Add `<%= error_message %>` template support for onerror fallback queries.
+
 
 1.30.1 - 2016-06-23
 -------------------